+++ app/src/main/AndroidManifest.xml
@@ -12,8 +12,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Messaging.MessagingBeta" />
-        <activity android:name=".Login.Login" />
+        <activity android:name=".View.Messaging.View" />
+        <activity android:name=".View.Login.View" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
