+++ app/src/main/java/com/example/messenger/MainActivity.java
@@ -3,11 +3,9 @@ package com.example.messenger;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 
-import com.example.messenger.Login.Login;
-import com.example.messenger.Messaging.MessagingBeta;
+import com.example.messenger.View.Messaging.View;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -23,18 +21,18 @@ public class MainActivity extends AppCompatActivity {
         btnLogin= (Button) findViewById(R.id.btnLogin);
         btnMessaging= (Button) findViewById(R.id.btnMessaging);
 
-        btnLogin.setOnClickListener(new View.OnClickListener() {
+        btnLogin.setOnClickListener(new android.view.View.OnClickListener() {
             @Override
-            public void onClick(View v) {
-                Intent doSomeThing= new Intent(MainActivity.this, Login.class);
+            public void onClick(android.view.View v) {
+                Intent doSomeThing= new Intent(MainActivity.this, com.example.messenger.View.Login.View.class);
                 startActivity(doSomeThing);
             }
         });
 
-        btnMessaging.setOnClickListener(new View.OnClickListener() {
+        btnMessaging.setOnClickListener(new android.view.View.OnClickListener() {
             @Override
-            public void onClick(View v) {
-                Intent doSomeThing= new Intent(MainActivity.this, MessagingBeta.class);
+            public void onClick(android.view.View v) {
+                Intent doSomeThing= new Intent(MainActivity.this, View.class);
                 startActivity(doSomeThing);
             }
         });
