+++ app/src/main/java/com/example/messenger/Presenter/Login/IPresenter.java
@@ -1,0 +1,6 @@
+package com.example.messenger.Presenter.Login;
+
+public interface IPresenter {
+    void Authenticate(String username, String password);
+    void CreateAccount(String username, String password);
+}
