+++ app/src/main/java/com/example/messenger/Presenter/Messaging/IPresenter.java
@@ -1,0 +1,4 @@
+package com.example.messenger.Presenter.Messaging;
+
+public interface IPresenter {
+}
