+++ app/src/main/java/com/example/messenger/Presenter/Messaging/Presenter.java
@@ -1,0 +1,4 @@
+package com.example.messenger.Presenter.Messaging;
+
+public class Presenter implements IPresenter {
+}
