+++ app/src/main/java/com/example/messenger/View/Login/IView.java
@@ -1,0 +1,6 @@
+package com.example.messenger.View.Login;
+
+public interface IView {
+    void successful();
+    void decline();
+}
