+++ app/src/main/java/com/example/messenger/View/Login/View.java
@@ -1,0 +1,75 @@
+package com.example.messenger.View.Login;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.widget.Button;
+import android.widget.EditText;
+
+import com.example.messenger.Presenter.Login.Presenter;
+import com.example.messenger.R;
+import com.example.messenger.Utils.RetrofitRoute;
+import com.example.messenger.Utils.RetrofitUtils;
+
+
+public class View extends AppCompatActivity implements IView {
+    EditText etUsername;
+    EditText etPassword;
+    Button btnSignIn;
+    Button btnSignUp;
+    RetrofitRoute retrofitRoute;
+    Presenter presenter = new Presenter(this);
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+
+        initVariable();
+
+        btnSignIn.setOnClickListener(new android.view.View.OnClickListener() {
+            @Override
+            public void onClick(android.view.View v) {
+                SignIn();
+            }
+        });
+
+        btnSignUp.setOnClickListener(new android.view.View.OnClickListener() {
+            @Override
+            public void onClick(android.view.View v) {
+                SignUp();
+            }
+        });
+    }
+
+    public void initVariable() {
+        etUsername = (EditText) findViewById(R.id.etUsername);
+        etPassword = (EditText) findViewById(R.id.etPassword);
+        btnSignIn = (Button) findViewById(R.id.btnSignin);
+        btnSignUp = (Button) findViewById(R.id.btnSignup);
+        retrofitRoute = RetrofitUtils.createRetrofitRoute();
+    }
+
+    public void SignIn() {
+        final String username = etUsername.getText().toString();
+        final String password = etPassword.getText().toString();
+        presenter.Authenticate(username, password);
+
+    }
+
+    public void SignUp() {
+        final String username = etUsername.getText().toString();
+        final String password = etPassword.getText().toString();
+        presenter.CreateAccount(username, password);
+    }
+
+    @Override
+    public void successful() {
+
+    }
+
+    @Override
+    public void decline() {
+
+    }
+}
