+++ app/src/main/res/layout/activity_messaging_beta.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Messaging.MessagingBeta">
+    tools:context=".View.Messaging.View">
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerview_message_list"
         android:layout_width="413dp"
